+++ HISTORY.rst
@@ -284,6 +284,7 @@
 * Add support for patchin dependency files using rewriting rules.
 * Allow searching documents by UUID.
 * Improve search negation logic.
+* Add support for search field transformations.
 
 3.1.11 (2019-04-XX)
 ===================